Missed changes

+++ USAble_DB/Tables/OrderMenuItems.sql
@@ -2,7 +2,7 @@
 (
 	[OrderId] INT NOT NULL,
 	[MenuItemId] INT NOT NULL,
-	[Quantity] SMALLINT NOT NULL,
+	[Quantity] INT NOT NULL,
 	PRIMARY KEY (OrderId, MenuItemId),
 	FOREIGN KEY (OrderId) REFERENCES [dbo].[Orders](Id),
 	FOREIGN KEY (MenuItemId) REFERENCES [dbo].[MenuItems](Id)
